+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.2.0 - 2016-03-09
+* **+/-** switched to `glob` to include the `baseDependencies` instead of search through the whole project
+	* **Please take a look at the changed defaults!**
+
 ## v0.1.0 - 2016-03-08
 * **+** added CHANGELOG
 * **+/-** changed README